+++ CHANGELOG.md
@@ -19,11 +19,31 @@ All notable changes to this project will be documented in this file.
 
 ## [منتشر نشده][unreleased]
 
+### [ایجاد / اضافه شد][#]
+
+- ایجاد توابع clone & instance
+
+### [تغییر کرد][#]
+
+-
+
+### [رفع / درست شد][#]
+
 -
 
 </div>
 
 ## [Unreleased]
+
+### [Added][#]
+
+- feat: added clone & instance function
+
+### [Changed][#]
+
+-
+
+### [Fixed][#]
 
 -
 
@@ -31,10 +51,80 @@ All notable changes to this project will be documented in this file.
 
 <div style="text-align: center;">
 
-## [1.0.0] - SH 1400-12-11 - G 2022-03-02
+## [1.2.4] - SH 1401-02-14 - G 2022-05-04
+
+</div>
+<div dir="rtl" style="text-align: right;">
+
+### [رفع / درست شد][#]
+
+- رفع خطای محاسباتی تاریخ setDate, setUTCDate
 
 </div>
 
+### [Fixed][#]
+
+- Fixed set day in setDate, setUTCDate
+
+#
+
+<div style="text-align: center;">
+
+## [1.2.0] - SH 1400-02-12 - G 2022-05-02
+
+</div>
+<div dir="rtl" style="text-align: right;">
+
+### [ایجاد / اضافه شد][#]
+
+- اضافه شدن مستندات
+- اضاقه شدن داکرایز
+- ایجاد تابع چک کننده زبان
+
+### [تغییر کرد][#]
+
+- تغییر نام تابع از setTimeServer به setTimeServerDiff
+- تغییر نام تابع از getTimeServer به getTimeServerDiff
+
+</div>
+
+### [Added][#]
+
+- Added documentation
+- Added dockerize
+- added check language function
+
+### [Changed][#]
+
+- Change setTimeServer function to setTimeServerDiff
+- Change getTimeServer function to getTimeServerDiff
+
+<div style="text-align: center;">
+
+#
+
+## [1.1.0] - SH 1400-02-09 - G 2022-04-29
+
+</div>
+<div dir="rtl" style="text-align: right;">
+
+### [ایجاد / اضافه شد][#]
+
+- اضافه شدن سمت کلاینت برای مرورگر
+
+</div>
+
+### [Added][#]
+
+- Added client side for browser
+
+<div style="text-align: center;">
+
+#
+
+## [1.0.0] - SH 1400-12-11 - G 2022-03-02
+
+</div>
 <div dir="rtl" style="text-align: right;">
 
 - انتشار نسخه اولیه
@@ -45,9 +135,11 @@ All notable changes to this project will be documented in this file.
 
 #
 
-[unreleased]: https://github.com/md-akhi/SHDateTime-js/compare/1.0.0...HEAD
-[2.0.0]: https://github.com/md-akhi/SHDateTime-js/compare/1.0.0...2.0.0
-[1.0.0]: https://github.com/md-akhi/SHDateTime-js/releases/tag/1.0.0
+[unreleased]: https://github.com/md-akhi/SHDateTime-js/compare/1.2.4...HEAD
+[1.2.4]: https://github.com/md-akhi/SHDateTime-js/compare/1.2.0...1.2.4
+[1.2.0]: https://github.com/md-akhi/SHDateTime-js/compare/1.1.1...1.2.0
+[1.1.0]: https://github.com/md-akhi/SHDateTime-js/compare/1.0.0-rc1...1.1.0
+[1.0.0]: https://github.com/md-akhi/SHDateTime-js/releases/tag/1.0.0-rc1
 [#]: #
 
 <!--
@@ -57,7 +149,6 @@ All notable changes to this project will be documented in this file.
 ## [x.y.z] - SH 4Y-2M-2D - G 4Y-2M-2D
 
 </div>
-
 <div dir="rtl" style="text-align: right;">
 
 -